Merge in STRATEGY/strategy-master from kaloyan to master

* commit 'f71d9b9c6316540352b4e83bdc1ba9891bee8c5a':
  STRATEGY-49 - migrations fix
  STRATEGY-49 - strategicheski-dokumenti.blade.php keep url when pagination

+++ database/migrations/2023_11_14_132340_add_strategic_act_link_to_strategic_document_table.php
@@ -13,9 +13,11 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('strategic_document', function (Blueprint $table) {
-            $table->string('strategic_act_link')->nullable()->after('strategic_act_number');
-        });
+        if (!Schema::hasColumn('strategic_document', 'strategic_act_link')) {
+            Schema::table('strategic_document', function (Blueprint $table) {
+                $table->string('strategic_act_link')->nullable()->after('strategic_act_number');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('strategic_document', function (Blueprint $table) {
-            $table->dropColumn('strategic_act_link');
-        });
+        if (Schema::hasColumn('strategic_document', 'strategic_act_link')) {
+            Schema::table('strategic_document', function (Blueprint $table) {
+                $table->dropColumn('strategic_act_link');
+            });
+        }
     }
 };
